requirements.txt
Django>=2
psycopg2-binary
djangorestframework
bcrypt
scrapy
setuptools
mysqlclient